requirements.txt
alembic==1.4.3
appdirs==1.4.4
certbot-py==0.10.1
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
distlib==0.3.1
dominate==2.6.0
filelock==3.0.12
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
pipenv==2020.11.15
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.25.0
six==1.15.0
SQLAlchemy==1.3.20
urllib3==1.26.2
virtualenv==20.1.0
virtualenv-clone==0.5.4
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
